prevent metric updates to be created in timelines that do not exist (#58)

+++ pkg/value/timeline/element/join.go
@@ -3,5 +3,5 @@ package element
 import "fmt"
 
 func Join(uni float64, nam string) string {
-	return fmt.Sprintf("%f,%s", uni, nam)
+	return fmt.Sprintf("%d,%s", int64(uni), nam)
 }
